Move ProjectChangedEvent and SceneChangedEvent to editor-commons
