+++ editor/src/main/com/mbrlabs/mundus/editor/ui/modules/menu/SceneMenu.kt
@@ -31,10 +31,10 @@ import com.mbrlabs.mundus.editor.Mundus
 import com.mbrlabs.mundus.editor.core.io.IOManager
 import com.mbrlabs.mundus.editor.core.io.IOManagerProvider
 import com.mbrlabs.mundus.editor.core.project.ProjectManager
-import com.mbrlabs.mundus.editor.events.ProjectChangedEvent
 import com.mbrlabs.mundus.editor.events.SceneAddedEvent
 import com.mbrlabs.mundus.editor.ui.UI
 import com.mbrlabs.mundus.editor.utils.Log
+import com.mbrlabs.mundus.editorcommons.events.ProjectChangedEvent
 
 /**
  * @author Marcus Brummer
